Refactor(header): export index 삭제
